refactor(headless): keep combobox changes out of autocomplete
