+++ packages/headless/src/primitives/autocomplete/autocomplete-root.tsx
@@ -57,11 +57,7 @@ function AutocompleteInner(props: AutocompleteProps) {
   const [selectedValue, setSelectedValue] = useControllableState<string | undefined>(
     props.value,
     props.defaultValue,
-    value => {
-      if (value !== undefined) {
-        props.onValueChange?.(value);
-      }
-    },
+    props.onValueChange as ((value: string | undefined) => void) | undefined,
   );
 
   const [activeIndex, setActiveIndex] = useState<number | null>(null);
@@ -72,7 +68,6 @@ function AutocompleteInner(props: AutocompleteProps) {
   const labelsRef = useRef<Array<string | null>>([]);
   const arrowRef = useRef<SVGSVGElement | null>(null);
   const popupRef = useRef<HTMLDivElement | null>(null);
-  const triggerRef = useRef<HTMLButtonElement | null>(null);
   const valuesByIndexRef = useRef<Map<number, string>>(new Map());
   const registerSelectedIndex = useCallback(
     (index: number, value: string) => {
@@ -130,13 +125,7 @@ function AutocompleteInner(props: AutocompleteProps) {
 
   const dismiss = useDismiss(floatingContext, {
     escapeKey: !inlineMode,
-    outsidePress(event) {
-      if (inlineMode) {
-        return false;
-      }
-      const target = event.target;
-      return !(target instanceof Node && triggerRef.current?.contains(target));
-    },
+    outsidePress: !inlineMode,
     bubbles: {
       escapeKey: inlineMode,
       outsidePress: inlineMode,
@@ -154,15 +143,6 @@ function AutocompleteInner(props: AutocompleteProps) {
   });
 
   const { getReferenceProps, getFloatingProps, getItemProps } = useInteractions([dismiss, role, listNav]);
-  const referenceProps = getReferenceProps();
-  const popupId = typeof referenceProps['aria-controls'] === 'string' ? referenceProps['aria-controls'] : undefined;
-
-  const focusInput = useCallback(() => {
-    const input = refs.domReference.current;
-    if (input instanceof HTMLElement) {
-      input.focus();
-    }
-  }, [refs.domReference]);
 
   const handleSelect = useCallback(
     (value: string, index: number, label: string) => {
@@ -206,15 +186,11 @@ function AutocompleteInner(props: AutocompleteProps) {
       elementsRef,
       labelsRef,
       popupRef,
-      triggerRef,
       arrowRef,
       valuesByIndexRef,
       setInlineMode,
       handleSelect,
       handleInputChange,
-      setOpen,
-      focusInput,
-      popupId,
       registerSelectedIndex,
       mounted,
       transitionProps,
@@ -234,9 +210,6 @@ function AutocompleteInner(props: AutocompleteProps) {
       selectedIndex,
       handleSelect,
       handleInputChange,
-      setOpen,
-      focusInput,
-      popupId,
       registerSelectedIndex,
       mounted,
       transitionProps,
